requirements.txt
Flask
sqlalchemy
python-dotenv
pymysql
Flask-Mail